Split test_skills.py into 4 per-concern files (#4611)
Mechanical refactor per plan - no production code changes, no test assertion
changes. Each test continues to execute from its new home and the source
file is deleted.

- test_skills_resolver.py: TestSkillResolver (minus 3 @-mention guards) +
  cache tests + TestSkillExecutionRoleParsing + TestSkillInfoSchemaExhaustiveness
  + activate_deps test. Owns shared module-level fixtures and constants
  (INTERNAL_SKILLS, ARCH_LENS_NAMES, EXP_LENS_NAMES, AUDIT_SKILL_NAMES,
  _get_bundled_skill_names, _all_skill_roots).
- test_skills_mention_guard.py: 3 @-mention structural guard tests as module
  level functions (no behavior in this file that would justify a class).
- test_skills_categories.py: TestSkillCategories + research-skill category
  tests + audit-skill-md existence tests + 5 tier-validation fall-through
  tests. Constants ARCH_LENS_NAMES / EXP_LENS_NAMES inlined to avoid
  cross-test-file import (pyright rejects cross-file imports under tests).
- test_skills_projection.py: EffectiveSkillCatalog visibility tests +
  projection tests + direct-install projection tests. _resolver_with_visibility_skills
  helper duplicated locally per plan.

Closes #4611
